getContent(): Add IDs to items retrieved from Dropbox
The Dropbox api provides IDs for all items requested. Up to now the app was ignoring these IDs since it didn't need them for anything. However, they are a helpful thing to save, especially for files.

When caching the streamable links for songs is implemented these IDs will be used as the key to store/retrieve by.

Every ID is a string. The first 3 characters are sliced because Dropbox starts all IDs with 'id:'. The string is reversed because empirical evidence has shown that, for some reason, Dropbox begins most IDs with the same ~19 characters.