app.js: Add updateTempLink()
The streamable links provided by Dropbox expire 4 hours after they are generated.

updateTempLink() is fired from updatePlaylist() when the user adds a song containing an expired streamable link to their playlist. updateTempLink() sends the song's path to the Dropbox api, receives a new streamable link, and uses the provided song id to update the link in the cached song object.