updatePlaylist(): Check cache for songs & validate link
updatePlaylist was requesting a new streamable link from the Dropbox api whenever a song was added to the playlist. Although a fine naive implementation, in the long run, this is not sustainable. It caused unnecessary api requests and unnecessary waiting for the user.

Streamable links are good for 4 hours after they are generated by Dropbox. Beatbox should cache streamable links when it receives them and, if the same song is requested again, should validate and serve data from the cache.

This commit implements checking the cache and, if the song is found, validating its streamable link as well as updating the streamable link if it is expired.